+++ src/Component/Sidebar.tsx
@@ -1,0 +1,52 @@
+import React from "react";
+
+import { makeStyles, useTheme, createMuiTheme } from "@material-ui/core/styles";
+import Drawer from "@material-ui/core/Drawer";
+import List from "@material-ui/core/List";
+import ListItem from "@material-ui/core/ListItem";
+import ListItemText from "@material-ui/core/ListItemText";
+
+const drawerWidth = 300;
+
+const useStyles = makeStyles((theme) => ({
+  // necessary for content to be below app bar
+  toolbar: theme.mixins.toolbar,
+  drawer: {
+    width: drawerWidth,
+    flexShrink: 0,
+  },
+  drawerPaper: {
+    width: drawerWidth,
+  },
+  drawerContainer: {
+    overflow: "auto",
+  },
+}));
+
+const Sidebar = () => {
+  const classes = useStyles();
+
+  return (
+    <Drawer
+      className={classes.drawer}
+      variant="permanent"
+      classes={{
+        paper: classes.drawerPaper,
+      }}
+    >
+      <div className={classes.toolbar} />
+      <div className={classes.drawerContainer}>
+        <List>
+          <ListItem button>
+            <ListItemText primary={"Home"} />
+          </ListItem>
+          <ListItem button>
+            <ListItemText primary={"Todo List"} />
+          </ListItem>
+        </List>
+      </div>
+    </Drawer>
+  );
+};
+
+export default Sidebar;
