--- src/Container/Sidebar.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Sidebar = () => {
-  return <div />;
-};
-
-export default Sidebar;
